Added minor fix
