+++ src/components/charts/ChartGroupUpdate.tsx
@@ -17,7 +17,7 @@ import ChartHeaderFilters from './ChartHeaderFilters'
 import { QueryParams } from './charts.util'
 import ChartEmptyState from '../common/emptyState/ChartEmptyState'
 import PageHeader from '../common/header/PageHeader'
-import {DetectBottom} from "../app/details/cIDetails/CIDetails";
+import DetectBottom from '../common/DetectBottom'
 
 export default function ChartGroupUpdate({}) {
     const history = useHistory()
