+++ src/components/common/index.ts
@@ -50,3 +50,4 @@ export * from './DatePickers/DayPickerRangeController'
 export * from './helpers/compareVersion'
 export * from './header/PageHeader'
 export * from './error/ErrorBar'
+export * from './DetectBottom'
